pyproject.toml
[tool.poetry]
name = "docassist"
version = "0.1.0"
description = ""
authors = ["Akhil Anilkumar <akhil.anilkumar@cubettech.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.0"
streamlit = "^1.45.1"
openai = "^1.78.1"
pymupdf = "^1.25.5"
pillow = "^11.2.1"
python-docx = "^1.1.2"
pdf2image = "^1.17.0"
pydicom = "^3.0.1"
langchain = "^0.3.25"
langchain-community = "^0.3.24"
bcrypt = "^4.3.0"
pypdf = "^5.6.0"
chromadb = "^1.0.12"
faiss-cpu = "^1.11.0"
pypdf2 = "^3.0.1"
pytesseract = "^0.3.13"
tiktoken = "^0.9.0"
langchain-openai = "^0.3.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
